test(chain): move CG-deposit failover suite to a focused sibling (#1535 review)
Address otReviewAgent: don't grow the 4.6k-line umbrella evm-adapter.unit.test.ts.
Move the `createOnChainContextGraph deposit read fails over` suite into a
dedicated `evm-adapter-cg-deposit-failover.unit.test.ts` (fast unit lane, no
Hardhat — the existing evm-adapter-cg-deposit.test.ts is real-Hardhat and the
wrong home for bare-provider doubles). Behavior + assertions unchanged; the
tiny per-file fixtures (minimalConfig/recorder/tooLowAllowanceRevert) are
copied, matching the repo's per-file test convention.
