+++ packages/chain/test/evm-adapter.unit.test.ts
@@ -4541,124 +4541,3 @@ describe('populateAndSignV10WithAllowanceRecovery — shared publish/update reco
     expect(signSpy.calls).toEqual([]);
   });
 });
-
-// #1340: the CG-register TooLowAllowance recovery must read the registration
-// deposit through the RPC-failover facade (`this.readContract`), not a bare
-// call on the signer's primary-bound `parametersStorage` handle. Before the
-// fix, a broken primary threw the read → it was swallowed to 0n → the TRAC
-// approve + retry never ran, so a new CG's first publish failed permanently
-// under a degraded primary despite healthy backups. These drive the REAL
-// readContract failover through `createOnChainContextGraph` (bare-provider
-// doubles, mirroring the C6 pattern above).
-describe('createOnChainContextGraph — deposit read fails over under a broken primary (#1340)', () => {
-  const DEPOSIT = 100n;
-  const CG_PARAMS = { accessPolicy: 0, publishPolicy: 1 } as const;
-
-  // Stub the CG-register path EXCEPT the deposit read + recovery control flow.
-  // `depositRead(i)` is the per-endpoint deposit view (i=0 primary, i=1 backup):
-  // readContract rebinds `parametersStorage.connect(provider)` per endpoint, so
-  // the double branches on provider identity to model a broken primary.
-  function makeCgAdapter(opts: {
-    depositRead: (endpointIndex: 0 | 1) => Promise<bigint>;
-    firstRevert?: () => unknown; // the first-attempt submitCreate error (default TooLowAllowance)
-  }) {
-    const a: any = new EVMChainAdapter(minimalConfig());
-    a.initialized = true;
-    a.init = async () => { a.initialized = true; };
-    // Isolate the failover read from the chainId preflight (not under test).
-    a.ensureConfiguredStaticChainIdValidated = async () => {};
-
-    const p0 = {}; const p1 = {};
-    a.providers = [p0, p1]; // two endpoints so readContract can fail over
-    a.rpcUrls = ['https://primary.example', 'https://backup.example'];
-
-    a.contracts = {
-      parametersStorage: {
-        connect: (p: unknown) => ({
-          contextGraphRegistrationDeposit: () => opts.depositRead(p === p0 ? 0 : 1),
-        }),
-      },
-      contextGraphs: { getAddress: async () => V10_KA_ADDRESS },
-      contextGraphStorage: {
-        interface: { parseLog: () => ({ name: 'ContextGraphCreated', args: { contextGraphId: 7n } }) },
-      },
-    };
-
-    const ensureSpy = recorder(async (..._a: unknown[]) => {});
-    a.ensureV10ApproveTrac = ensureSpy;
-
-    // submitCreate → sendContractTransaction: revert once, then mine a receipt.
-    let sends = 0;
-    const receipt = { logs: [{ topics: [], data: '0x' }], hash: '0xhash', blockNumber: 1, index: 0, status: 1 };
-    const sendSpy = recorder(async (..._a: unknown[]) => {
-      sends += 1;
-      if (sends === 1) throw (opts.firstRevert ?? tooLowAllowanceRevert)();
-      return receipt;
-    });
-    a.sendContractTransaction = sendSpy;
-
-    return { a, ensureSpy, sendSpy };
-  }
-
-  const retryable429 = () => { const e = new Error('429 too many requests'); (e as any).status = 429; return e; };
-
-  it('cure: broken primary → deposit read fails over to the backup → exactly one approve → create succeeds', async () => {
-    let primaryHits = 0; let backupHits = 0;
-    const { a, ensureSpy, sendSpy } = makeCgAdapter({
-      depositRead: async (i) => {
-        if (i === 0) { primaryHits += 1; throw retryable429(); }
-        backupHits += 1; return DEPOSIT;
-      },
-    });
-
-    const result = await a.createOnChainContextGraph(CG_PARAMS);
-
-    // Failover actually happened (not just "approve fired"):
-    expect(primaryHits).toBe(1);            // primary view tried and threw
-    expect(backupHits).toBe(1);             // failed over to the backup
-    // Approve fired exactly once, with the real (failed-over) deposit, force=true:
-    expect(ensureSpy.calls).toHaveLength(1);
-    expect(ensureSpy.calls[0][2]).toBe(DEPOSIT);
-    expect(ensureSpy.calls[0][4]).toBe(true);
-    // Recovery is one-shot: initial submit reverts, retry mines.
-    expect(sendSpy.calls).toHaveLength(2);
-    expect(result.success).toBe(true);
-    expect(result.contextGraphId).toBe(7n);
-  });
-
-  it('preserved semantics: deposit read throws on ALL endpoints → no approve, original TooLowAllowance propagates', async () => {
-    const { a, ensureSpy, sendSpy } = makeCgAdapter({ depositRead: async () => { throw retryable429(); } });
-    await expect(a.createOnChainContextGraph(CG_PARAMS)).rejects.toThrow('TooLowAllowance');
-    expect(ensureSpy.calls).toEqual([]); // conservative: never approve on an unreadable deposit
-    expect(sendSpy.calls).toHaveLength(1); // only the initial submit; no retry
-  });
-
-  it('healthy primary unchanged: deposit answers on the primary, backup never consulted, one approve', async () => {
-    let backupHits = 0;
-    const { a, ensureSpy } = makeCgAdapter({
-      depositRead: async (i) => { if (i === 1) backupHits += 1; return DEPOSIT; },
-    });
-    const result = await a.createOnChainContextGraph(CG_PARAMS);
-    expect(backupHits).toBe(0); // no failover on a healthy primary
-    expect(ensureSpy.calls).toHaveLength(1);
-    expect(result.success).toBe(true);
-  });
-
-  it('dormant deposit (reads 0) → no approve (deposit===0n branch), original revert propagates', async () => {
-    const { a, ensureSpy } = makeCgAdapter({ depositRead: async () => 0n });
-    await expect(a.createOnChainContextGraph(CG_PARAMS)).rejects.toThrow('TooLowAllowance');
-    expect(ensureSpy.calls).toEqual([]);
-  });
-
-  it('guard intact: a non-TooLowAllowance first revert → deposit NOT read, no approve', async () => {
-    let reads = 0;
-    const { a, ensureSpy, sendSpy } = makeCgAdapter({
-      depositRead: async () => { reads += 1; return DEPOSIT; },
-      firstRevert: () => new Error('execution reverted: SomeOtherError()'),
-    });
-    await expect(a.createOnChainContextGraph(CG_PARAMS)).rejects.toThrow('SomeOtherError');
-    expect(reads).toBe(0); // guard short-circuits before the deposit read
-    expect(ensureSpy.calls).toEqual([]);
-    expect(sendSpy.calls).toHaveLength(1);
-  });
-});
